refactor: ErrorPage 파일명 변경, 매직넘버 처리
