+++ src/store/MovieStore.ts
@@ -1,4 +1,4 @@
-import ErrorPage from '../components/ErrorPage';
+import UserNotifyPage from '../components/UserNotifyPage';
 import { Movie } from '../index.d';
 
 const options = {
@@ -29,13 +29,13 @@ class MovieStore {
     )
       .then(async (response) => {
         if (!response.ok) {
-          new ErrorPage(String(response.status)).renderError();
+          new UserNotifyPage(String(response.status)).renderError();
         }
         const responseJSON = await response.json();
 
         // 검색 결과가 없는 경우
         if (String(response.status)[0] === '2' && responseJSON.results.length === 0) {
-          new ErrorPage(String(response.status)).renderError();
+          new UserNotifyPage(String(response.status)).renderError();
         }
         return responseJSON;
       })
